@@ -17,59 +17,56 @@
 	"bugs": {
 		"url": "https://github.com/CTAPbIuMABP/goat/issues"
 	},
-	"license": {
-		"type": "MIT",
-		"url": "https://github.com/CTAPbIuMABP/goat/blob/master/MIT-LICENSE.txt"
-	},
+	"license": "MIT",
 	"dependencies": {
 		"bcrypt-nodejs": "0.0.3",
-		"body-parser": "1.12.4",
-		"bootstrap": "3.3.4",
+		"body-parser": "1.13.1",
+		"bootstrap": "3.3.5",
 		"cookie-parser": "1.3.5",
 		"connect-mongo": "0.8.1",
 		"classnames": "2.1.2",
-		"css-loader": "0.14.4",
+		"css-loader": "0.15.1",
 		"cuid": "1.2.5",
 		"debug": "2.2.0",
-		"eslint": "0.21.0",
-		"eslint-loader": "0.11.2",
-		"eslint-plugin-react": "2.3.0",
-		"express": "4.12.3",
-		"express-session": "1.10.4",
-		"extract-text-webpack-plugin": "0.5.0",
+		"eslint": "0.23.0",
+		"eslint-loader": "0.14.0",
+		"eslint-plugin-react": "2.5.0",
+		"express": "4.12.4",
+		"express-session": "1.11.3",
+		"extract-text-webpack-plugin": "0.8.1",
 		"file-loader": "0.8.4",
 		"flux": "2.0.3",
-		"font-awesome": "4.2.0",
-		"jquery": "2.1.3",
+		"font-awesome": "4.3.0",
+		"jquery": "2.1.4",
 		"less-loader": "2.2.0",
 		"lodash": "3.9.3",
 		"moment": "2.10.3",
-		"mongodb": "2.0.33",
-		"mongoose": "4.0.1",
-		"morgan": "1.5.3",
-		"nodemailer": "1.3.4",
+		"mongodb": "2.0.35",
+		"mongoose": "4.0.6",
+		"morgan": "1.6.0",
+		"nodemailer": "1.4.0",
 		"passport": "0.2.2",
 		"passport-local": "1.0.0",
 		"passport-facebook": "1.0.3",
 		"passport-google-oauth": "0.1.5",
 		"q": "1.4.1",
-		"react": "0.13.1",
+		"react": "0.13.3",
 		"react-router": "0.13.3",
-		"react-bootstrap": "0.22.6",
-		"react-router-bootstrap": "0.15.0",
-		"react-hot-loader": "1.2.4",
-		"style-loader": "0.8.2",
-		"webpack": "1.7.3",
-		"webpack-dev-server": "1.8.0"
+		"react-bootstrap": "0.23.7",
+		"react-hot-loader": "1.2.7",
+		"react-router-bootstrap": "0.16.0",
+		"style-loader": "0.12.3",
+		"webpack": "1.10.0",
+		"webpack-dev-server": "1.10.1"
 	},
 	"devDependencies": {
 		"mocha": "2.2.1",
 		"nodemon": "1.3.7",
-		"babel": "5.4.7",
-		"babel-runtime": "5.4.7",
-		"babel-core": "5.4.7",
-		"babel-loader": "5.1.3",
-		"babel-eslint": "3.1.8"
+		"babel": "5.6.14",
+		"babel-core": "5.6.15",
+		"babel-runtime": "5.6.15",
+		"babel-loader": "5.1.4",
+		"babel-eslint": "3.1.19"
 	},
 	"engines": {
 		"node": "v0.12.x"
